Add code to demonstrate the type() method for data types in Python
This commit adds code to demonstrate the usage of the type() method in Python. The code showcases different data types and their corresponding output when passed to the type() method.

The code includes examples for the following data types:

>>Integer: The variable "a" is assigned the value 20, and the type() method is used to determine its data type.
>>Floating-point number: The variable "b" is assigned the value 3.145, and the type() method is used to determine its data type.
>>String: The variable "c" is assigned the string "Hello World", and the type() method is used to determine its data type.
String representation of a number: The variable "d" is assigned the string "7.14", and the type() method is used to determine its data type.
>>List: The variable "lst" is assigned a list containing the elements [2, 3, 5, 8, 7, 8], and the type() method is used to determine its data type.
>>Tuple: The variable "tpl" is assigned a tuple containing the elements (7, 8, 2, 4), and the type() method is used to determine its data type.

By including this code, we provide a clear demonstration of how to use the type() method to find the data type of different variables in Python. This understanding of data types is essential for effective programming and data manipulation.