Create  Working with Sets in Python,   Set Comprehension, Set Differences, and Set Methods
The code demonstrates various operations and methods related to sets in Python. It starts by creating sets using the set() method and curly braces. It shows how to add and remove elements from a set using the add() and remove() methods, check if an element is present in a set using the "in" keyword, and perform set operations such as union, intersection, and difference. The code also includes examples of set comprehension, counting occurrences of elements in a set, clearing a set, and using various set methods like isdisjoint(), issuperset(), and issubset().