Create Python Date Object with Examples
This code demonstrates how to use the time and datetime modules in Python to work with dates and times. The time module provides functions for getting the current time, converting between different representations of time, and suspending execution for a specified amount of time. The datetime module provides classes for representing dates, times, and datetime objects.

The code first imports the time and datetime modules. It then defines a few variables to store the current time, the current date, and a custom date. It then uses the time module to get the current time in seconds since the epoch, as a string, and the current year. It then uses the sleep function to suspend execution for 10 seconds. Finally, it uses the datetime module to create a custom date object, get the current date, get the current month, create a custom time object, get the current datetime object, create a timedelta object, and perform a timedelta calculation.

This code can be used to learn how to use the time and datetime modules in Python to work with dates and times.