Create Object - Oriented Programming in Python
This code demonstrates the basics of object-oriented programming (OOP) in Python. It defines a class called Person, which has two data members: name and age. It also defines a member function called display_info, which prints the values of the name and age data members. Finally, it creates an object of the Person class and assigns values to the data members